+++ NEWS.md
@@ -4,3 +4,7 @@
 # DoddLabRawMS 0.1.1
 * Added a `NEWS.md` file to track changes to the package.
 * Added a `README` file
+
+# DoddLabRawMS 0.1.11
+* Added a transponson parameter set
+* Mofidy the function to skip rt correction error